+++ tests/testthat/test-write.R
@@ -6,18 +6,18 @@ tbl_iris <- get_tbl_iris()
 test_that("write_csv_chunkwise works", {
   tmp <- tempfile()
   write_csv_chunkwise(tbl_iris, tmp, row.names = FALSE)
-  iris2 <- read.csv(tmp)
+  iris2 <- read.csv(tmp, stringsAsFactors = TRUE)
   expect_equal(iris2, iris)
 })
 
 test_that("write_csv2_chunkwise works", {
   tmp <- tempfile()
   write_csv2_chunkwise(tbl_iris, tmp, row.names = FALSE)
-  iris2 <- read.csv2(tmp)
+  iris2 <- read.csv2(tmp, stringsAsFactors = TRUE)
   expect_equal(iris2, iris)
 
   write_chunkwise(tbl_iris, tmp, format="csv2", row.names = FALSE)
-  iris2 <- read.csv2(tmp)
+  iris2 <- read.csv2(tmp, stringsAsFactors = TRUE)
   expect_equal(iris2, iris)
 })
 
@@ -27,7 +27,7 @@ test_that("write_chunkwise to db works", {
   expect_equal(as.data.frame(iris2)[1:4], as.data.frame(tbl_iris)[1:4])
   tmp <- tempfile()
   write_chunkwise(iris2, tmp, row.names = FALSE)
-  iris3 <- read.csv(tmp)
+  iris3 <- read.csv(tmp, stringsAsFactors = TRUE)
   expect_equal(iris3, iris)
 })