+++ NEWS.md
@@ -1,3 +1,9 @@
+Version 0.3.2
+
+* Fix for change of default in stringsAsFactors for R 4.0
+
+* Added `stringsAsFactors` as argument to `read_chunkwise` functions.
+
 Version 0.3.1
 
 * Fix for dplyr upgrade from 0.5 to 0.6